fixed, moved exported functions that arent components into separate files (#495)
